// src/lib/api/models/TypeEnum.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * * `booking_confirmation` - Booking Confirmation
 * * `booking_request` - Booking Request
 * * `booking_cancelled` - Booking Cancelled
 * * `message` - Message
 * * `review` - Review
 * * `payment_confirmation` - Payment Confirmation
 * * `payment_failed` - Payment Failed
 * * `property_approved` - Property Approved
 * * `property_rejected` - Property Rejected
 */
export enum TypeEnum {
    BOOKING_CONFIRMATION = 'booking_confirmation',
    BOOKING_REQUEST = 'booking_request',
    BOOKING_CANCELLED = 'booking_cancelled',
    MESSAGE = 'message',
    REVIEW = 'review',
    PAYMENT_CONFIRMATION = 'payment_confirmation',
    PAYMENT_FAILED = 'payment_failed',
    PROPERTY_APPROVED = 'property_approved',
    PROPERTY_REJECTED = 'property_rejected',
}